Update icon path, add Bitfocus in name and lowercase Companion.xml
The icon URL was returning a 404. Updated the URL to work again. When this first came out on Unraid I put a post up about it on the official Companion Facebook user group and one of the lead developers said that the package should be named "Bitfocus Companion". So I've updated that as well.
I also changed it from Companion.xml to companion.xml so that it sorts alphabetically in the templates folder in Github.