Update mealiev1.xml
Update Mealie Template for the new images are hosted on ghcr.io

Also, updated the ports, since before their were 2 ports, now just one. That is also why I have 9000 default to 3000 and 3000 was previously the default web port, and want to avoid breaking changes if possible. I tested it on myself and it achieved the desired result.

Also updated variables under advanced for postgres as optional, and cleaned up the description as well.